refactor: 운동방 목록 조회 및 생성 API 연동
